Refactor: Kept only filter logic for creating chart data
