+++ apps/balancer-tools/src/app/apr/pool/(components)/HistoricalCharts/HistoricalData/HistoricalVolumeChart.tsx
@@ -1,0 +1,27 @@
+import { redDarkA } from "@radix-ui/colors";
+import { PlotType } from "plotly.js";
+
+import { PoolStatsResults } from "#/app/apr/api/route";
+
+import { generateAndTrimAprCords } from "..";
+
+export default function HistoricalVolumeChartData(
+  apiResult: PoolStatsResults,
+): Plotly.Data {
+  const HOVERTEMPLATE = "$%{y:.2f}";
+
+  const trimmedTotalAprData = generateAndTrimAprCords(
+    apiResult.perRound,
+    (result) => result.volume,
+    0,
+  );
+
+  return {
+    name: "Volume",
+    hovertemplate: HOVERTEMPLATE,
+    x: trimmedTotalAprData.x,
+    y: trimmedTotalAprData.y,
+    line: { shape: "spline", color: redDarkA.redA9 } as const,
+    type: "scatter" as PlotType,
+  };
+}
